pyproject.toml
[project]
name = "mc-backup"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = [
    "deptry>=0.21.0",
    "mypy>=1.13.0",
    "pytest>=8.3.3",
    "ruff>=0.7.3",
]


[tool.mypy]
plugins = []

strict = true
follow_imports = "normal"
ignore_missing_imports = false

allow_redefinition = false
disallow_any_explicit = false
ignore_errors = false
local_partial_types = true
no_implicit_optional = true
strict_optional = true
warn_no_return = true
warn_return_any = true
warn_unreachable = true

pretty = true
show_column_numbers = true
show_error_codes = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true

[tool.ruff]
src = ["src", "tests"]
[tool.ruff.lint]
fixable = [
    "F",
    "E",
    "W",
    "I",
    "COM",
    "UP",
    "RUF",
]
unfixable = [
    "F841", # Variable  is assigned to but never used
]
select = ["ALL"]
ignore = [
    "E501", # Line Length
    "COM812",  # ruff format conflict
    "ISC001",  # ruff format conflict
    "D10", # Mandatory Docstrings
    "D203",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"src/app/alembic/*" = ["INP001"]
"src/app/alembic/versions/*" = ["D400", "D415"]
"src/app/db/models/*" = ["A003"]
"src/app/*/schema.py" = ["A003"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.mccabe]
max-complexity = 6

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = []

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod"]
staticmethod-decorators = ["staticmethod"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"
